requirements.txt
fastapi==0.116.1
openai==1.99.1
pandas==2.3.1
pinecone-client==3.0.0
protobuf==6.31.1
python-dotenv==1.1.1
Requests==2.32.4
streamlit==1.47.1
tqdm==4.67.1
uvicorn==0.24.0
